move TestRouterSendConfigUpdateToConsenterStub from the config_dissemination_test.go  to router_test.go
